requirements.txt
simplecv # Definitely not easy to install
Pillow
svgwrite
redis
